fix(payments+tickets): decode treasury amount-as-string; skip stock reservation for event tickets
- treasury client: response amount fields use a string-or-number `money` type — fixes
  "payment gateway is unavailable" caused by `cannot unmarshal string into float64`, so orders
  now return a usable payment intent + initiate_url.
- checkout: event-ticket lines (metadata.is_ticket) are excluded from inventory stock reservation
  (SERVICE items have no balance → was 409), and a best-effort event-capacity pre-check rejects
  sold-out tiers before payment.

+++ internal/modules/payments/repository_treasury.go
@@ -135,7 +135,7 @@ func treasuryIntentToPaymentIntent(resp *treasury.PaymentIntentResponse) *Paymen
 		ProviderIntentID: resp.ProviderIntentID,
 		ClientSecret:     resp.ClientSecret,
 		Status:           PaymentIntentStatus(resp.Status),
-		Amount:           resp.Amount,
+		Amount:           float64(resp.Amount),
 		Currency:         resp.Currency,
 		CreatedAt:        resp.CreatedAt,
 		UpdatedAt:        resp.CreatedAt,
